dt-bindings: net: dsa: nxp,sja1105: improve MDIO bus bindings
The SJA1110 switch uses the mdios property for its two MDIO buses. Instead
of a pattern, define two mdio nodes. This ensures the same compatible
string won't be used twice.

Move the comment to the description of mdios, mdio@0, and mdio@1
properties. Disallow the mdios property for SJA1105. Require at least one
of the MDIO buses to be defined to prevent empty mdios child node.

Signed-off-by: Arınç ÜNAL <arinc.unal@arinc9.com>